# server/openapi.yaml
openapi: 3.1.0
info:
  title: Aria API
  version: 2.0.0
  description: |
    Aria v2 music-server API. Paths and JSON shapes are byte-compatible with the
    legacy Node server; v2 adds GET /healthz, GET /api/events (SSE), and optional
    limit/offset pagination on GET /api/tracks.

    ID scheme: track `id` = sha1 hex of the file path relative to MUSIC_DIR;
    `albumId` = sha1 hex of lowercase(album directory + "\0" + album title), the
    directory being the file's folder with a CD1/Disc 2 subfolder folded into its
    parent, and the album artist replacing the (empty) directory at library root.

    Error convention: validation failures are 400 with `{"error": "<reason>"}`;
    not-found is a bare 404 (text/plain body "Not Found", no JSON);
    upstream MusicBrainz failures are 502 with `{"error": "..."}`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/enrich/people:
    post:
      tags: [enrichment]
      summary: Warm faces/bios for visible names
      description: Fire-and-forget enrichment of up to 50 names (each < 200 chars); non-string or too-long entries are silently dropped.
      operationId: warmPeople
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                names:
                  type: array
                  maxItems: 50
                  items: { type: string, maxLength: 199 }
      responses:
        '200':
          description: How many names were newly queued.
          content:
            application/json:
              schema:
                type: object
                required: [queued]
                properties:
                  queued: { type: integer }
...